Rename and move file. Add private constructor.
