+++ src/main/java/org/jabref/gui/push/PushToVim.java
@@ -11,9 +11,9 @@ import org.jabref.gui.JabRefExecutorService;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.icon.JabRefIcon;
 import org.jabref.logic.l10n.Localization;
-import org.jabref.model.JabRefStringConstants;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
+import org.jabref.model.push.PushToApplicationConstants;
 import org.jabref.preferences.PreferencesService;
 import org.jabref.preferences.PushToApplicationPreferences;
 
@@ -22,7 +22,7 @@ import org.slf4j.LoggerFactory;
 
 public class PushToVim extends AbstractPushToApplication implements PushToApplication {
 
-    public static final String NAME = JabRefStringConstants.VIM;
+    public static final String NAME = PushToApplicationConstants.VIM;
 
     private static final Logger LOGGER = LoggerFactory.getLogger(PushToVim.class);
