+++ src/main/java/org/jabref/gui/push/PushToWinEdt.java
@@ -3,12 +3,12 @@ package org.jabref.gui.push;
 import org.jabref.gui.DialogService;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.icon.JabRefIcon;
-import org.jabref.model.JabRefStringConstants;
+import org.jabref.model.push.PushToApplicationConstants;
 import org.jabref.preferences.PreferencesService;
 
 public class PushToWinEdt extends AbstractPushToApplication implements PushToApplication {
 
-    public static final String NAME = JabRefStringConstants.WIN_EDT;
+    public static final String NAME = PushToApplicationConstants.WIN_EDT;
 
     public PushToWinEdt(DialogService dialogService, PreferencesService preferencesService) {
         super(dialogService, preferencesService);
