+++ src/main/java/org/jabref/model/push/PushToApplicationConstants.java
@@ -1,12 +1,15 @@
-package org.jabref.model;
+package org.jabref.model.push;
 
-public class JabRefStringConstants {
+public class PushToApplicationConstants {
 
-    // PushTo name constants
     public static final String EMACS = "Emacs";
     public static final String LYX = "LyX/Kile";
     public static final String TEXMAKER = "Texmaker";
     public static final String TEXSTUDIO = "TeXstudio";
     public static final String VIM = "Vim";
     public static final String WIN_EDT = "WinEdt";
+
+    private PushToApplicationConstants() {
+    }
+
 }
