+++ src/main/java/org/jabref/preferences/JabRefPreferences.java
@@ -95,7 +95,6 @@ import org.jabref.logic.util.Version;
 import org.jabref.logic.util.io.AutoLinkPreferences;
 import org.jabref.logic.util.io.FileHistory;
 import org.jabref.logic.xmp.XmpPreferences;
-import org.jabref.model.JabRefStringConstants;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.BibEntryType;
 import org.jabref.model.entry.BibEntryTypesManager;
@@ -106,6 +105,7 @@ import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.types.EntryType;
 import org.jabref.model.entry.types.EntryTypeFactory;
 import org.jabref.model.metadata.SaveOrderConfig;
+import org.jabref.model.push.PushToApplicationConstants;
 import org.jabref.model.strings.StringUtil;
 
 import org.slf4j.Logger;
@@ -1741,12 +1741,12 @@ public class JabRefPreferences implements PreferencesService {
     @Override
     public PushToApplicationPreferences getPushToApplicationPreferences() {
         Map<String, String> applicationCommands = new HashMap<>();
-        applicationCommands.put(JabRefStringConstants.EMACS, get(PUSH_EMACS_PATH));
-        applicationCommands.put(JabRefStringConstants.LYX, get(PUSH_LYXPIPE));
-        applicationCommands.put(JabRefStringConstants.TEXMAKER, get(PUSH_TEXMAKER_PATH));
-        applicationCommands.put(JabRefStringConstants.TEXSTUDIO, get(PUSH_TEXSTUDIO_PATH));
-        applicationCommands.put(JabRefStringConstants.VIM, get(PUSH_VIM));
-        applicationCommands.put(JabRefStringConstants.WIN_EDT, get(PUSH_WINEDT_PATH));
+        applicationCommands.put(PushToApplicationConstants.EMACS, get(PUSH_EMACS_PATH));
+        applicationCommands.put(PushToApplicationConstants.LYX, get(PUSH_LYXPIPE));
+        applicationCommands.put(PushToApplicationConstants.TEXMAKER, get(PUSH_TEXMAKER_PATH));
+        applicationCommands.put(PushToApplicationConstants.TEXSTUDIO, get(PUSH_TEXSTUDIO_PATH));
+        applicationCommands.put(PushToApplicationConstants.VIM, get(PUSH_VIM));
+        applicationCommands.put(PushToApplicationConstants.WIN_EDT, get(PUSH_WINEDT_PATH));
 
         return new PushToApplicationPreferences(
                 applicationCommands,
@@ -1757,12 +1757,12 @@ public class JabRefPreferences implements PreferencesService {
 
     @Override
     public void storePushToApplicationPreferences(PushToApplicationPreferences preferences) {
-        put(PUSH_EMACS_PATH, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.EMACS));
-        put(PUSH_LYXPIPE, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.LYX));
-        put(PUSH_TEXMAKER_PATH, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.TEXMAKER));
-        put(PUSH_TEXSTUDIO_PATH, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.TEXSTUDIO));
-        put(PUSH_VIM, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.VIM));
-        put(PUSH_WINEDT_PATH, preferences.getPushToApplicationCommandPaths().get(JabRefStringConstants.WIN_EDT));
+        put(PUSH_EMACS_PATH, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.EMACS));
+        put(PUSH_LYXPIPE, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.LYX));
+        put(PUSH_TEXMAKER_PATH, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.TEXMAKER));
+        put(PUSH_TEXSTUDIO_PATH, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.TEXSTUDIO));
+        put(PUSH_VIM, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.VIM));
+        put(PUSH_WINEDT_PATH, preferences.getPushToApplicationCommandPaths().get(PushToApplicationConstants.WIN_EDT));
 
         put(PUSH_EMACS_ADDITIONAL_PARAMETERS, preferences.getEmacsArguments());
         put(PUSH_VIM_SERVER, preferences.getVimServer());
